Reviewed by John.
        - Fixed a few crashes resulting from NULL parameters to JSClassCreate.

        * API/JSClassRef.cpp:
        (JSClassCreate):
        (JSClassRelease):
        * API/testapi.c: Added test for NULL parameters.
        (main):



Canonical link: https://commits.webkit.org/12801@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@15213 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ JavaScriptCore/API/JSClassRef.cpp
@@ -37,7 +37,7 @@ JSClassRef JSClassCreate(JSContextRef, JSStaticValue* staticValues, JSStaticFunc
     JSClassRef jsClass = new __JSClass;
     if (staticValues) {
         jsClass->staticValues = new __JSClass::StaticValuesTable();
-        while (staticValues->name != 0) {
+        while (staticValues->name) {
             jsClass->staticValues->add(Identifier(staticValues->name).ustring().rep(), 
                                        new StaticValueEntry(staticValues->getProperty, staticValues->setProperty, staticValues->attributes));
             ++staticValues;
@@ -46,14 +46,18 @@ JSClassRef JSClassCreate(JSContextRef, JSStaticValue* staticValues, JSStaticFunc
     
     if (staticFunctions) {
         jsClass->staticFunctions = new __JSClass::StaticFunctionsTable();
-        while (staticFunctions->name != 0) {
+        while (staticFunctions->name) {
             jsClass->staticFunctions->add(Identifier(staticFunctions->name).ustring().rep(), 
                                           new StaticFunctionEntry(staticFunctions->callAsFunction, staticFunctions->attributes));
             ++staticFunctions;
         }
     }
     
-    jsClass->callbacks = *callbacks;
+    if (callbacks)
+        jsClass->callbacks = *callbacks;
+    else
+        jsClass->callbacks = kJSObjectCallbacksNone;
+    
     jsClass->parent = parentClass;
     
     return JSClassRetain(jsClass);
@@ -68,11 +72,15 @@ JSClassRef JSClassRetain(JSClassRef jsClass)
 void JSClassRelease(JSClassRef jsClass)
 {
     if (--jsClass->refCount == 0) {
-        deleteAllValues(*jsClass->staticValues);
-        delete jsClass->staticValues;
+        if (jsClass->staticValues) {
+            deleteAllValues(*jsClass->staticValues);
+            delete jsClass->staticValues;
+        }
 
-        deleteAllValues(*jsClass->staticFunctions);
-        delete jsClass->staticFunctions;
+        if (jsClass->staticFunctions) {
+            deleteAllValues(*jsClass->staticFunctions);
+            delete jsClass->staticFunctions;
+        }
 
         delete jsClass;
     }
